python/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
async-timeout==4.0.2
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
distlib==0.3.6
entrypoints==0.4
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.29
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
multidict==6.0.4
mysqlclient==2.1.1
numpy==1.23.5
openai==0.26.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pipenv==2022.10.12
platformdirs==2.5.2
protobuf==3.20.3
py-mon==1.1.0
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.42
streamlit==1.15.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
twilio==7.16.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchdog==2.2.0
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
